requirements.txt

flask
openai
pillow
